portplow/requirements.txt
Django==1.9.3
apache-libcloud==0.20.1
celery==3.1.23
chaussette==1.3.0
circus==0.13.0
django-bootstrap3==7.0.1
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-filter==0.12.0
django-memoize==1.3.1
django-redis==4.4.2
django-session-security==2.3.2
djangorestframework==3.3.2
paramiko==1.16.0
psutil==4.1.0
psycopg2==2.6.1
git+https://github.com/dlitz/pycrypto.git@v2.7a1
# pycrypto==2.7a1
python-libnmap==0.7.0